stereoanywhere/requirements.txt
albumentations
cmapy
kornia
matplotlib
numba
numpy
opencv_contrib_python
opencv_python
opt_einsum
Pillow
scikit_learn
timm==1.0.15
torch==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm
xformers==0.0.27.post2
pyrealsense2
pupil-apriltags
